Hoping this fixed it.  Stupid bugs!
git-svn-id: file:///home/oestrich/tmp/img/trunk@95 ea04b438-0a61-0410-93d0-9bc465a3bab0

+++ queue.php
@@ -25,11 +25,11 @@ if(isset($_POST)){
 	$lncln->dequeue($_POST);
 }
 
-if(isset($_GET['delete'])){
+if(isset($_GET['delete']) && $lncln->isAdmin){
 	$deletion = $lncln->delete($_GET['delete']);
 }
 
-if(isset($_GET['obscene']) && $isLoggedIn){
+if(isset($_GET['obscene']) && $lncln->isAdmin){
 	$obscene = $lncln->obscene($_GET['obscene']);
 }
